backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# HTTP client and API integration
httpx==0.25.2
requests==2.31.0

# Data processing and file handling
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2
PyPDF2==3.0.1

# AI and OpenAI integration
openai==1.3.7

# Async and concurrency
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Type hints and development
typing-extensions==4.8.0
pydantic==2.5.0

# Date and time handling
python-dateutil==2.8.2

# JSON and data serialization
orjson==3.9.10

# Database (SQLite)
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0